pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "career-compass"
description= "Backend for Career Compass"
version = "2024.8.0"
dependencies = [
  "fastapi[standard]",
  "alembic",
  "sqlalchemy",
  "pyjwt",
  "bcrypt", 
  "pydantic-settings",
  "python-dotenv",
  "psycopg2-binary",
  "httpx"
]
requires-python = ">= 3.12"
authors = [
  {name = "Michael Hall", email= "michael@rubberduck-labs.com"}
]
maintainers = [
  {name= "Michael Hall", email = "michael@rubberduck-labs.com"}
]
readme = "README.md"
classifiers = [
  "Development Status :: 1 - Planning",
  "Framework :: FastAPI",
  "Programming Language :: Python :: 3.12"
]


[project.urls]
Homepage = "https://example.com"
Documentation = "https://readthedocs.org"
Repository = "https://github.com/me/spam.git"
Issues = "https://github.com/me/spam/issues"
Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/careercompass"]

[tool.hatch.envs.default]
type = "virtual"
path = ".venv"
dependencies = ["pyright", "pytest", "pytest-cov"]

[tool.hatch.envs.default.env-vars]
SQLALCHEMY_DATABASE_URL = "sqlite:///./foo.db"

[tool.hatch.envs.default.scripts]
dev = "fastapi dev main.py"
update-db = "alembic upgrade head"
start-app = "fastapi run app/main.py"